+++ CHANGES.rst
@@ -11,5 +11,7 @@ To be released.
 - Added :mod:`nirum.datastructures` module and
   :class:`~nirum.datastructures.Map` which is an immutable dictionary.
   [:issue:`66`]
+- Added :class:`nirum.datastructures.List` which is an immutable list.
+  [:issue:`49`]
 
 __ https://github.com/spoqa/nirum/blob/f1629787f45fef17eeab8b4f030c34580e0446b8/docs/serialization.md